requirements.txt
python-dotenv==1.0.0
psycopg2
flask==2.3.3
flask-cors==4.0.0
reportlab==4.0.4
markdown2==2.4.12
weasyprint==60.2
pylatex==1.4.1
requests==2.31.0